NetworkProfiler class split in two different classes for better readability
