+++ src/yarpconnectionsinfo/main.cpp
@@ -304,4 +304,3 @@ digraph G {
     b3 -> d3 [label="tcp" color=blue];
 }
 */
-
